Two different fixes: Better package target (guava) and split apart one test that'd fail if run in a grouping
