+++ src/test/java/org/terasology/logic/health/RestorationTest.java
@@ -15,7 +15,7 @@
  */
 package org.terasology.logic.health;
 
-import com.google.api.client.util.Sets;
+import com.google.common.collect.Sets;
 import org.junit.Before;
 import org.junit.Test;
 import org.terasology.entitySystem.entity.EntityManager;
